feat: move players in game folder
