+++ src/components/game/players/blend.tsx
