+++ src/components/game/players/index.tsx
